Improved OOD. Reduced repetitive code. Added more config options
Before, a reference to the statusbar was being passed into each app, and then they
would use this reference to add themselves to the statusbar. Changed the code so that
the statusbar adds each app. The StatusItem / StatusSingle and StatusMulti shared
functions. Moved the repeated code to StatusWidget. Can now search for more items in
the config file.