Improved Atlas code style and formatting, as well as directory structure.
Atlas applications are no longer template heavy. A class template for the
AtlasApp class is still used, as well as for the "attach" function, but no other
templates are utilized in the function definitions.

The static modifier was put in place more frequently, in parts of the code where
multiple calls were made to a function, so repeating certain memory allocations
would be wasteful.

Improved upon the structure of the enumerated alignment and signal
types. Encapsulating them in a struct is preferable, thus the code was changed
to reflect this.

Code was restructured in a directory hierarchy that follows the "app", "core",
"common" style.