Updated the AtlasApple class so that the Workspace application is now working.
The Workspace application is the trickiest one since it uses a Label array with
a focus element; it is now working though. Many static functions from AtlasApple
will be moved to a generic namespace containing all the necessary
functions. That way, the class is not cluttered with too many functions.