Moved code back to an earlier design structure.
Makes use of more coding and completely removes the configuration file part.

Renamed files so everything is neat and lowercase.

Revised headers of each file so hopefully there's no errors there.

Made code more modular and contained, as much as I could at least.

Hopefully made things make more sense and more intuitive. Oh, and there's a
window manager section too. It's mainly for the dwm window manager, but it
essentially attempts to inject as little code (and thus work for the user) as
possible into dwm.c and it's respective config.h.

Along that same vein, there's an atom section which handles all the _NET_* atoms
that are necessary. So far it's really just _NET_CURRENT_DESKTOP and
_NET_NUMBER_OF_DESKTOPS that is needed.