Added window manager functions to clean up interface as well as logging.
The status bar to window manager interface can be initialized in the window
manager and intialization also includes handling several kill signals, as Atlas
is forked and run in the background. The signal handling for kill signals is to
ensure that the child process is terminated when the parent dies.

Logging in strategic places in the code was also added, in case an error occurs
and I need to go back and check where it happended.