Separated interface further and switched type of signal container.
Moved the memory mapped file functions to its own source file (from ipc).

Edited types so that they are defined in the headers they will be used
in. Avoids having to make up some dummy temporary name that will never be used.

Changed type of signal container to a std::map so that the key can be the signal
key, and the value can be the Atlas application. When the signal handler is
called, all I need to do is search for the key (instead of looping) and I can
easily call the callback.